added typescript configuration to api folder
