+++ api/.eslintrc.json
@@ -1,0 +1,16 @@
+{
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {
+        "ecmaVersion": 2018,
+        "sourceType": "module"
+    },
+    "extends": ["plugin:@typescript-eslint/recommended"],
+    "env": {
+        "node": true
+    },
+    "rules": {
+        "indent": "off",
+        "@typescript-eslint/indent": "off",
+        "@typescript-eslint/explicit-function-return-type": "off"
+    }
+}
