+++ api/.gitignore
@@ -2,6 +2,8 @@
 
 # dependencies
 /node_modules
+/api/node_modules
+/app/node_modules
 /.pnp
 .pnp.js
