--- api/db.js
@@ -1,14 +1,0 @@
-const mongoose = require("mongoose");
-require("dotenv").config();
-
-mongoose
-  .connect(process.env.DB_URL, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true
-  })
-  .then(() => {
-    console.log("DB connected");
-  })
-  .catch(err => {
-    console.error("Error connecting to Mongo", err);
-  });
