+++ api/db.ts
@@ -1,0 +1,16 @@
+import mongoose from "mongoose";
+import dotenv from "dotenv";
+
+dotenv.config();
+
+mongoose
+  .connect(`${process.env.DB_URL}`, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true
+  })
+  .then(() => {
+    console.log("DB connected");
+  })
+  .catch(err => {
+    console.error("Error connecting to Mongo", err);
+  });
