+++ api/package.json
@@ -4,16 +4,28 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "dotenv": "^8.2.0",
+    "body-parser": "^1.19.0",
     "express": "^4.17.1",
     "mongodb": "^3.5.4",
-    "mongoose": "^5.9.2",
-    "node": "^13.8.0",
-    "nodemon": "^2.0.2"
+    "mongoose": "^5.9.2"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "@types/body-parser": "^1.19.0",
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.2",
+    "@types/mongodb": "^3.5.0",
+    "@types/mongoose": "^5.7.3",
+    "@types/node": "^13.7.7",
+    "@typescript-eslint/eslint-plugin": "^2.22.0",
+    "@typescript-eslint/parser": "^2.22.0",
+    "dotenv": "^8.2.0",
+    "eslint": "^6.8.0",
+    "nodemon": "^2.0.2",
+    "ts-node": "^8.6.2",
+    "typescript": "^3.8.3"
+  },
   "scripts": {
-    "start": "nodemon server.js"
+    "start": "nodemon server.ts"
   },
   "author": "",
   "license": "ISC"
