--- api/server.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const app = express();
-const port = 4000;
-
-app.listen(port, () => {
-  console.log("Server is running on ", port);
-});
