+++ api/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+        "target": "es6",
+        "module": "commonjs",
+        "rootDir": ".",
+        "outDir": "./build",
+        "esModuleInterop": true,
+        "strict": true
+    }
+}
